Make upgrade work easily and automatically
