+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Existing commands are backward compatible with V1 datasets, however some new fun
     - Text output: Features are now labelled as `<dataset>:feature:<primary_key>`, consistent with meta items that are labelled as `<dataset>:meta:<meta_item_name>`
     - JSON output also uses "feature" and "meta" as keys for the different types of changes, instead of "featureChanges" and "metaChanges".
  * `sno show -o json` header key changed to `sno.show/v1`, which is not an applyable patch. Use `sno create-patch` to create a patch.
+ * `sno upgrade` now only takes two arguments: `sno upgrade [source] [dest]`. No other arguments are required or accepted, exactly how to upgrade the repository is detected automatically. 
 
 ### Other changes in this release
