+++ tests/test_upgrade.py
@@ -7,18 +7,16 @@ H = pytest.helpers.helpers()
 
 @pytest.mark.slow
 @pytest.mark.parametrize(
-    "archive,layer",
+    "archive",
     [
-        pytest.param("points0.snow", H.POINTS.LAYER, id="points"),
-        pytest.param("polygons0.snow", H.POLYGONS.LAYER, id="polygons-pk"),
-        pytest.param("table0.snow", H.TABLE.LAYER, id="table"),
+        pytest.param("points0.snow", id="points"),
+        pytest.param("polygons0.snow", id="polygons-pk"),
+        pytest.param("table0.snow", id="table"),
     ],
 )
-def test_upgrade_00_02(archive, layer, data_archive, cli_runner, tmp_path, chdir):
+def test_upgrade_v0(archive, data_archive, cli_runner, tmp_path, chdir):
     with data_archive(archive) as source_path:
-        r = cli_runner.invoke(
-            ["upgrade", "00-02", source_path, tmp_path / "dest", layer]
-        )
+        r = cli_runner.invoke(["upgrade", source_path, tmp_path / "dest"])
         assert r.exit_code == 0, r
         assert r.stdout.splitlines()[-1] == "Upgrade complete"
 
@@ -26,15 +24,15 @@ def test_upgrade_00_02(archive, layer, data_archive, cli_runner, tmp_path, chdir
         r = cli_runner.invoke(["log"])
         assert r.exit_code == 0, r
 
-        if layer == H.POINTS.LAYER:
+        if archive == "points0.snow":
             assert r.stdout.splitlines() == [
-                "commit a91ac9bf11cf6a07647fdc8700dbcfab95685804",
+                "commit 0c64d8211c072a08d5fc6e6fe898cbb59fc83d16",
                 "Author: Robert Coup <robert@coup.net.nz>",
                 "Date:   Thu Jun 20 15:28:33 2019 +0100",
                 "",
                 "    Improve naming on Coromandel East coast",
                 "",
-                "commit 4c2fabf6c58ecfede957f2205caa4f85ac6b56ad",
+                "commit 7bc3b56f20d1559208bcf5bb56860dda6e190b70",
                 "Author: Robert Coup <robert@coup.net.nz>",
                 "Date:   Tue Jun 11 12:03:58 2019 +0100",
                 "",
@@ -51,13 +49,13 @@ def test_upgrade_00_02(archive, layer, data_archive, cli_runner, tmp_path, chdir
         pytest.param("table", H.TABLE.LAYER, id="table"),
     ],
 )
-def test_upgrade_02_05(archive, layer, data_archive, cli_runner, tmp_path, chdir):
+def test_upgrade_v1(archive, layer, data_archive, cli_runner, tmp_path, chdir):
     with data_archive(archive) as source_path:
         r = cli_runner.invoke(["status", "--output-format=json"])
         assert r.exit_code == 0, r
         src_branch = json.loads(r.stdout)["sno.status/v1"]["branch"]
 
-        r = cli_runner.invoke(["upgrade", "02-05", source_path, tmp_path / "dest"])
+        r = cli_runner.invoke(["upgrade", source_path, tmp_path / "dest"])
         assert r.exit_code == 0, r
         assert r.stdout.splitlines()[-1] == "Upgrade complete"
 
@@ -84,13 +82,3 @@ def test_upgrade_02_05(archive, layer, data_archive, cli_runner, tmp_path, chdir
         assert r.exit_code == 0, r
         dest_branch = json.loads(r.stdout)["sno.status/v1"]["branch"]
         assert dest_branch == src_branch
-
-
-def test_upgrade_list(cli_runner):
-    r = cli_runner.invoke(["upgrade"])
-    assert r.exit_code == 0, r
-    assert r.stdout.splitlines()[-3:] == [
-        "Commands:",
-        "  00-02  Upgrade a v0.0/v0.1 Sno repository to Sno v0.2",
-        "  02-05  Upgrade a v0.2/v0.3/v0.4 Sno repository to Sno v0.5",
-    ]
